adjust transaction queue processing gate
There seemed to be some issues with relying on asyncio queues for
processing transaction queues in that they seemed to be getting
stuck. This was probably due to an error occuring that prevented one
of the tasks from triggering the next in line and removing itself.

This changes things to use a redis key as the gate, with an expiry in
case something unexpected happens that causes the process not to free
up the key.

This also moves closer to allowing multiple instances of the manager
to run simultainiously.