process token transactions alongside base transaction
There was a bug here when running behind a load balancer where if
eth_getTransactionReceipt failed and triggered a retry of the
transaction update, the base transaction has already been set as
confirmed and thus the update wouldn't execute the 2nd time, leaving
the updated token balances unchecked.

This fix puts all the database updating side of things into the same
place so that the database update only happens when everything is
ready to be updated.